v0.0.2.4
Summary:

feat: implement timestamp-based recording IDs and friendly speaker naming

Description:

Backend & Database:
Switched Recording IDs from auto-increment integers to timestamp-based BigIntegers (YYYYMMDDHHMMSSmm) for better uniqueness.
Restricted ID length to 16 digits (10ms precision) to ensure compatibility with JavaScript's MAX_SAFE_INTEGER.
Updated all related models (Transcript, RecordingSpeaker, RecordingTag) to use BigInteger for foreign keys.
Renamed the User table to users to resolve conflicts with the PostgreSQL reserved keyword user.
Added pool_pre_ping=True to the synchronous database engine to prevent psycopg2.OperationalError in Celery workers after database restarts.
Worker:
Implemented logic to rename raw diarization labels (SPEAKER_00) to user-friendly sequential names (Speaker 1, Speaker 2) based on order of appearance.
Ensured identified Global Speakers retain their names.
Fixed a regex syntax warning in tasks.py.
Companion App (Rust):
Refactored the entire application to use i64 instead of i32 for recording_id to accommodate
the new timestamp-based format.
* Updated server routes, state management, and uploader logic to handle 64-bit integers.

+++ companion/src/audio.rs
@@ -70,7 +70,7 @@ pub fn run_audio_loop(state: Arc<AppState>, command_rx: Receiver<AudioCommand>) 
 }
 
 fn start_segment(
-    recording_id: i32,
+    recording_id: i64,
     sequence: i32,
     state: Arc<AppState>,
     is_recording: Arc<AtomicBool>
