+++ companion/src/state.rs
@@ -13,7 +13,7 @@ pub enum AppStatus {
 
 pub struct AppState {
     pub status: Mutex<AppStatus>,
-    pub current_recording_id: Mutex<Option<i32>>,
+    pub current_recording_id: Mutex<Option<i64>>,
     pub current_sequence: Mutex<i32>,
     pub audio_command_tx: Sender<AudioCommand>,
     pub config: Mutex<Config>,
@@ -24,7 +24,7 @@ pub struct AppState {
 
 #[derive(Debug, Clone)]
 pub enum AudioCommand {
-    Start(i32), // recording_id
+    Start(i64), // recording_id
     Pause,
     Resume,
     Stop,
